requirements.txt
streamlit>=1.28.0
google-generativeai>=0.8.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
